msm: camera: isp: Reset the allow callback flag in aeb exp config
Currently, the epoch which generates a callback to hw manager is
controlled by a flag which is set in exp order config. This flag
is not reset on subsequent exp configs, which can cause multiple
epoch callbacks. Fix this by resetting the flag on every update.
Also, add eof callback to the same functionality so that flash
trigger can be done at correct time for aeb on case.

CRs-Fixed: 3904425
Change-Id: Iedde758957c70464da5d4d6d634ae442315c63b8
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>