msm: camera: mem_mgr: camera IOVA region restrict to 36bits(64GB)
Camera HW we have support of max 36 bits and for that max address
range we have defined in dtsi is 64GB. New kernel IOVA address range
change to 512GB(39bits), this is changing the camera IOVA region mapping
and causing SMMU fault. This change will help to restrict the
IOVA address range to 36 bits.

CRs-Fixed: 3956403
Change-Id: Ie91fa6ea43e9c885ef2c295d9fc3d1d62c1ea731
Signed-off-by: Soumen Ghosh <quic_soumeng@quicinc.com>