msm: camera: isp: Add framework to override the CSID header settings
With SWI compatibility, Register headers for new targets should be able
to leverage the base target and override if any new feature has to added.
This commit adds support for the framework to override for compatible
targets.
During probe, the override function would be called which will set the
settings as per the current target while inheriting the base features.

CRs-Fixed: 3800048
Change-Id: I7b61a7eae170dac13926bab94ae19f44a0b0f344
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>