msm: camera: isp: Decoupled multi VC-DT fixes
This change handles the case where num VC > num DT.
For example VC0 -> DT 43 and VC1 -> DT -1.
Additional check is added for DT validity before
determining the decode format for that entry in the
array, as well as setting that DT to zero when it
comes time to program the field in the register.

CRs-Fixed: 3962703
Change-Id: I3c288fb97f29c5326e82102c97106c2341052044
Signed-off-by: Li Sha Lim <quic_lishlim@quicinc.com>