msm: camera: common: Add support for OPP framework
Unlike in the downstream kernel where the clock driver
takes care of configuring the voltage rail to the desired
level based on the clock rate set, we need to use OPP
framework and the APIs it supports to set the rate for the
main SRC clock of the device which internally takes care
of the voltage level as well. However, we are limiting the
support here to one OPP table for the device that corresponds
to the main SRC clock of the device. Other SRC clocks will
continue to use the clock driver API and the camera driver
takes care of scaling up the main SRC clock as required.

CRs-Fixed: 3981067
Change-Id: Iab383317b1f4d55e81e1c65b898d2d9b9a79ecfd
Signed-off-by: Vijay Kumar Tumati <quic_vtumati@quicinc.com>