mona: solve sensor probe fail when pmic error
details:
It's a workaround gerrit
An error in the PMIC will cause the sensor to fail to power on.
Even if the PMIC recovers normally later, the camera will not turn
on after the power on failure is determined.
So adding a retry operation after power on failure in the kernel to
avoid this issue.

Change-Id: Ia1a9100429bb2b21560cbee82d1b7b744bdeca6f
Signed-off-by: yangmf3 <yangmf3@motorola.com>
Reviewed-on: https://gerrit.mot.com/3178597
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Pdo Scm <hudsoncm@motorola.com>
Reviewed-by: wangdp11 <wangdp11@motorola.com>
Reviewed-by: Guan Xin <guanxin6@motorola.com>
Reviewed-by: Zhichao Chen <chenzc2@motorola.com>
Submit-Approved: Jira Key