msm: camera: common: handle undefined VMRM symbols with upstream kernel
Clean up the VMRM headers to not include some header files that
contain the external symbols which are not available on the
upstream kernel. However, when compiling with the downstream kernel,
where VMRM feature needs to be supported, SPECTRA_VMRM can be used
to enable the dependencies including external symbols.

CRs-Fixed: 4019268
Change-Id: Ifc9856929a483f1f671f6e71acc742e093de67d7
Signed-off-by: Vijay Kumar Tumati <quic_vtumati@quicinc.com>