msm: camera: tpg: Full register dump on CSID Fatal Error
Enable automatic logging of TPG register values upon fatal errors on
CSID or RX path to aid in debugging. Remove the CAM_IO_DUMP mask as it
is unnecessary to set a different mask for dumping other than that of
clients. Update the functions responsible for dumping to include
additional parameters for client mask and error case. Modify all client
calls to these functions accordingly.

CRs-Fixed: 3941022
Change-Id: I568f30a61632188c674bd60490b804f85c567677
Signed-off-by: Abhilash Mahapatra <quic_abhmah@quicinc.com>