msm: camera: isp: Fix the memory leak for init packet cmd buffer
In case of EPCR, refcount for first init packet is not getting put
from driver. For the next packet, cmd buffer ref count is put
as a part of flush or while handling buf_dones.
This commit fixes the leak by calling the put API for first init
packet while updating the settings for init packet.

CRs-Fixed: 4030704
Change-Id: I8677f09dcfc0c53dc8f351b6137e61d20c0ab5a7
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>