msm: camera: isp: Add support for dynamic offset computatation
With current design of TFE/IFE headers, offsets of all registers
in different modules are maintained relative to the core start address.
Register offsets within a module remain constant among different targets
while the base offset of the module changes relative to the core.
With current design, for any change in base offset of module, all
underlying offsets need to be changed.

This commit adds base address of different modules relative
to the core and maintains the  actual register offsets relative to the base
module. For any target, if the base of the module is changed, offsets of
internal registers need not be changed in the header file.
Updating the base address would help. This reduces the effort to
maintain the addresses of the registers for each target.

CRs-Fixed: 3800048
Change-Id: Ie46d5151ed7c2d975d6e0e5e6db7fa3a3537c0a5
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>