msm: camera: isp: Add dynamic bus register offsets computation support
Current TFE headers maintains the absolute offsets of the registers.
If the bus base offset is changed w.r.t to the core start, all the offsets
need to be changed. With the compatibility changes in place, bus base
offset can be overridden using the override function and the underlying
register offsets can be computed dynamically.
This commit adds support for common base of the bus and changes
the offsets of registers w.r.t the bus base.

CRs-Fixed: 3800048
Change-Id: I79ba9fb4ddfe7a69d03be57eeff999b0b50c5134
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>