msm: camera: isp: Handle per request regdump during skip frame
In ISP driver, the sequence followed for a request application to
buffers being consumed is as follows:

EPOCH: event is sent to CRM to apply pending request and ISP SW
after applying, moves the request to waiting request list.
REG_UPDATE: event is sent to ISP SW and it dequeues the request
from waiting request list and moves it to active request list.
BUF_DONE: event is sent to ISP SW about buffer consumption at the
bus port. Once buf dones for all ports are received for the
active request, the request is freed and sent to free request
list.

When per request regdump is enabled for ISP, ISP SW starts reading
the registers at the time of EPOCH and associates the read data to
the request active at that time. This makes sure that the register
dump is always associated with request that is currently active.

However, there can be a case when we do not send the apply request
to ISP due to any other request not yet available on the same link
for other devices. At this time a skip frame is encountered in CRM
and register dump is missed for the current active request as by
the time next EPOCH comes, all buf dones are handled and active
request list becomes empty.

To overcome such scenarios this change adds handling register dump
as part of skip frame notification received from CRM to ISP.

CRs-Fixed: 4033285
Change-Id: Ic230d7656a21600a52857c361796a1bb3958cb7b
Signed-off-by: Abhilash Kumar <quic_krabhi@quicinc.com>