msm: camera: isp: Restructure the WM client info in header files
Current design has redundant bookkeeping of output ports
and WM clients in the header files. This design is not
aligned with the HW forward compatibility. With target changes,
the WM client information changes which can be fetched from the query caps.
There would be redundant effort to maintain the outport information.
This commit adds support to get the out port type based on the
client index which would remain same with forward compatibility.

CRs-Fixed: 4106367
Change-Id: I8d87e54230b77f0a6f354159210b16d3237576bf
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>