msm: camera: csiphy: Refactor configuration to be lane-based
Currently, the driver programs the configuration for all CPHY/DPHY
lanes irrespective of if a lane is enabled or not. This poses
problems for more complex configurations such as CPHY + DPHY,
which are not in use currently.

To meet this future need, this change refactors the code to only
program the settings for each lane in use. This way, CPHY + DPHY
lanes can be muxed without issues. With this approach, a lot of
redundancies in the header files are cleaned up as well.

CRs-Fixed: 4044549
Change-Id: I1b90cd9ac77bf00b89e88eed69e325ff777b91dc
Signed-off-by: Li Sha Lim <quic_lishlim@quicinc.com>