msm: camera: csiphy: Update DPHY clk ln selection logic
Currently, the driver makes some hardcoded assumptions
based on the set of PHY lane combos we support.
However, this set is not complete, and this assumption
will fail for some valid combos.

This change makes the selection based off the physical
placing of the clock lanes, and how they can be configured.
By doing so, we cover all valid combinations.

CRs-Fixed: 4044549
Change-Id: I9a107ea30cd53bc1fad50bea3fae322dcc4f5f24
Signed-off-by: Li Sha Lim <quic_lishlim@quicinc.com>