msm: camera: isp: Update calculation logic to detect change in ife bw
Currently, bandwidth voting for ife is based on change in sum of
the camnoc bandwidth values for all applicable paths. In certain
cases, it is possible that mnoc bandwidth changes without any
change in camnoc. Update the detection logic to take the total
voted mnoc bandwidth into account as well.

CRs-Fixed: 4154384
Change-Id: I8f8db8e71ef0181d81404f1b04a6cf3127442210
Signed-off-by: Mukund Madhusudan Atre <matre@qti.qualcomm.com>