msm: camera: isp: Add support to trigger error on non faulting stream
In some cases, it is possible that non-faulting SW context or
stream needs to be recovered even if the originating hardware
is not associated to that stream. Add error code to notify this
case to user space. Add support to trigger error on non-faulting
stream to simulate this scenario.

CRs-Fixed: 4132618
Change-Id: I719c27a4b1ee559d764e3eeafe986411f3e14cd2
Signed-off-by: Mukund Madhusudan Atre <matre@qti.qualcomm.com>