msm: camera: cpas: Define all camnoc_info struct data in headers only
Instead of deriving regbase value, camnoc name for different NOCs
inside driver, set the regbase variable in header itself as this
is constant based on the type of NOC. Have camnoc names as global
string array and use it while printing with indexing of camnoc type.
Intention is to avoid updating hw info header data structures in
driver as much as possible.

CRs-Fixed: 4174291
Change-Id: I0e316b0d520abc79803a6a31f22da7d7aef6f786
Signed-off-by: Pavan Kumar Chilamkurthi <pchilamk@qti.qualcomm.com>